+++ src/components/containers/expander.styl
@@ -1,18 +1,30 @@
-.container
+.expander
+  transition: max-height 150ms ease-out
   position: relative
+  overflow: hidden
   
-.controlArea
-  width: 100%
-  padding-bottom: 12px
-  z-index: 1
-  button
-    z-index: 1
-  &:after
-    content: ''
-    pointer-events: none
+.expanderMask
+  position: absolute
+  top: 0
+  bottom: 0
+  left: 0
+  right: 0
+  display: flex
+  flex-direction: row
+  align-items: flex-end
+  background: linear-gradient(-180deg, rgba(255,255,255,0.00) 0%, #FFFFFF 100%);
+
+.expander-button
+  position: initial
+  margin-bottom: 2px
+  border-color: tertiary
+  &:hover
+    border-color: primary
+    box-shadow: 2px 2px 0 primary
+  &::before
     position: absolute
+    top: 0
     bottom: 0
     left: 0
-    width: 100%
-    height: 350px
-    background: linear-gradient(-180deg, rgba(255,255,255,0.00) 0%, #FFFFFF 100%)
+    right: 0
+    content: ""
